refactor plan campaing progression into FSD
