+++ src/fsd/1-pages/plan-campaign-progression/campaign-progression-header.tsx
@@ -155,5 +155,3 @@ export const CampaignProgressionHeader = () => {
         </span>
     );
 };
-
-export default CampaignProgressionHeader;
