+++ src/fsd/1-pages/plan-campaign-progression/campaign-progression-rankup-goals.tsx
@@ -3,41 +3,34 @@ import { ColDef, AllCommunityModule, themeBalham } from 'ag-grid-community';
 import { AgGridReact } from 'ag-grid-react';
 import React, { useState } from 'react';
 
-import { UnitShardIcon } from '@/fsd/5-shared/ui/icons';
-import { RarityIcon } from '@/fsd/5-shared/ui/icons/rarity.icon';
-import { StarsIcon } from '@/fsd/5-shared/ui/icons/stars.icon';
+import { Rank } from '@/fsd/5-shared/model';
+import { MiscIcon, UnitShardIcon } from '@/fsd/5-shared/ui/icons';
 
-import { CharactersService, ICharacterData } from '@/fsd/4-entities/character';
+import { CharactersService, ICharacterData, RankIcon } from '@/fsd/4-entities/character';
+import { ICharacterUpgradeMow, ICharacterUpgradeRankGoal } from '@/fsd/4-entities/goal';
 
-import { CampaignData } from 'src/v2/features/campaign-progression/campaign-progression.models';
-import { ICharacterUnlockGoal, ICharacterAscendGoal } from 'src/v2/features/goals/goals.models';
+import { CampaignData } from './campaign-progression.models';
 
 interface Props {
     campaignData: CampaignData;
-    goals: Array<ICharacterUnlockGoal | ICharacterAscendGoal>;
+    goals: Array<ICharacterUpgradeRankGoal | ICharacterUpgradeMow>;
 }
 
-export const CampaignProgressionAscensionGoals: React.FC<Props> = ({ campaignData, goals }) => {
+export const CampaignProgressionRankupGoals: React.FC<Props> = ({ campaignData, goals }) => {
+    const [goalDefs] = useState(getColumnDefs());
+
     /** @returns the goal with the given ID. */
-    function getGoal(goalId: string): ICharacterAscendGoal | ICharacterUnlockGoal | undefined {
-        const filtered: Array<ICharacterAscendGoal | ICharacterUnlockGoal> = goals.filter(
+    function getGoal(goalId: string): ICharacterUpgradeRankGoal | ICharacterUpgradeMow | undefined {
+        const filtered: Array<ICharacterUpgradeRankGoal | ICharacterUpgradeMow> = goals.filter(
             goal => goal.goalId == goalId
         );
-        if (filtered.length == 0) return undefined;
+        if (filtered.length == 0) {
+            return undefined;
+        }
         if (filtered.length > 1) {
             console.warn('multiple goals with ID ' + goalId + ' found.');
         }
         return filtered[0];
-    }
-
-    /**
-     * @returns the ascension goal with the given ID, or undefined if it either
-     *          doesn't exist, or isn't an ascension goal.
-     */
-    function getAscensionGoal(goalId: string): ICharacterAscendGoal | undefined {
-        const goal = getGoal(goalId);
-        if (goal && 'rarityStart' in goal && 'starsStart' in goal) return goal as ICharacterAscendGoal;
-        return undefined;
     }
 
     /**
@@ -47,10 +40,8 @@ export const CampaignProgressionAscensionGoals: React.FC<Props> = ({ campaignDat
     function getGoalData(campaignData: CampaignData): any[] {
         const rowData: any[] = [];
         for (const [goalId, cost] of campaignData[1].goalCost) {
-            const goal = getAscensionGoal(goalId);
-            if (goal) {
-                rowData.push({ goalData: [{ goalId: goalId, goalCost: cost }] });
-            }
+            const goal = getGoal(goalId);
+            if (goal) rowData.push({ goalData: [{ goalId: goalId, goalCost: cost }] });
         }
         return rowData;
     }
@@ -59,6 +50,47 @@ export const CampaignProgressionAscensionGoals: React.FC<Props> = ({ campaignDat
     function getGoalUnit(goalId: string): ICharacterData | undefined {
         if (!getGoal(goalId)) return undefined;
         return CharactersService.getUnit(getGoal(goalId)!.unitId) ?? undefined;
+    }
+
+    /**
+     * @returns the starting rank of the goal. If the
+     * goal is an unlock goal, returns 0, meaning 'Locked'.
+     */
+    function getGoalRankStart(goalId: string): number {
+        const goal = getGoal(goalId);
+        if (goal && 'rankStart' in goal) return goal.rankStart;
+        return 0;
+    }
+
+    /**
+     * @returns the starting rank of the goal. If the
+     * goal is an unlock goal, returns 1, meaning stone1.
+     */
+    function getGoalRankEnd(goalId: string): number {
+        // I have no idea why, but the typescript compiler in
+        // WebStorm thinks that we have to jump through all of
+        // these hoops here, but not in getGoalRankStart above.
+        const goal = getGoal(goalId);
+        let rankEnd: number = 1;
+        if (!goal) return 1;
+        Object.entries(goal).forEach(([key, value]) => {
+            if (key == 'rankEnd') rankEnd = value as number;
+        });
+        return rankEnd;
+    }
+
+    /** @returns the web link to the rank-up lookup of the given goal. */
+    function getRankLookupHref(goalId: string): string {
+        const rankStart = Math.max(getGoalRankStart(goalId), 1);
+        const rankEnd = getGoalRankEnd(goalId);
+        return (
+            '../../learn/rankLookup?character=' +
+            getGoalUnit(goalId)?.id +
+            '&rankStart=' +
+            Rank[rankStart] +
+            '&rankEnd=' +
+            Rank[Math.max(rankStart + 1, rankEnd)]
+        );
     }
 
     /**
@@ -75,11 +107,13 @@ export const CampaignProgressionAscensionGoals: React.FC<Props> = ({ campaignDat
                     if (!params.data.goalData || !params.data.goalData[0]) return '';
                     const goalData = params.data.goalData[0];
                     return (
-                        <UnitShardIcon
-                            icon={getGoalUnit(goalData.goalId)?.icon ?? '(undefined)'}
-                            height={30}
-                            tooltip={getGoalUnit(goalData.goalId)?.icon}
-                        />
+                        <a href={getRankLookupHref(goalData.goalId)}>
+                            <UnitShardIcon
+                                icon={getGoalUnit(goalData.goalId)?.icon ?? '(undefined)'}
+                                height={30}
+                                tooltip={getGoalUnit(goalData.goalId)?.icon}
+                            />
+                        </a>
                     );
                 },
             },
@@ -90,23 +124,11 @@ export const CampaignProgressionAscensionGoals: React.FC<Props> = ({ campaignDat
                 cellRenderer: (params: any) => {
                     if (!params.data.goalData || !params.data.goalData[0]) return '';
                     const goalData = params.data.goalData[0];
-                    const goal: ICharacterAscendGoal = getAscensionGoal(goalData.goalId)!;
-                    return <RarityIcon rarity={goal.rarityStart} />;
+                    return <RankIcon rank={getGoalRankStart(goalData.goalId)} />;
                 },
             },
             {
                 headerName: 'C',
-                width: 60,
-                cellStyle: { align: 'center' },
-                cellRenderer: (params: any) => {
-                    if (!params.data.goalData || !params.data.goalData[0]) return '';
-                    const goalData = params.data.goalData[0];
-                    const goal: ICharacterAscendGoal = getAscensionGoal(goalData.goalId)!;
-                    return <StarsIcon stars={goal.starsStart} />;
-                },
-            },
-            {
-                headerName: 'D',
                 width: 35,
                 cellStyle: { align: 'center' },
                 cellRenderer: (params: any) => {
@@ -115,31 +137,29 @@ export const CampaignProgressionAscensionGoals: React.FC<Props> = ({ campaignDat
                 },
             },
             {
-                headerName: 'E',
+                headerName: 'D',
                 width: 60,
                 cellStyle: { align: 'center' },
                 cellRenderer: (params: any) => {
                     if (!params.data.goalData || !params.data.goalData[0]) return '';
                     const goalData = params.data.goalData[0];
-                    const goal: ICharacterAscendGoal = getAscensionGoal(goalData.goalId)!;
-                    return <RarityIcon rarity={goal.rarityEnd} />;
+                    return <RankIcon rank={getGoalRankEnd(goalData.goalId)} />;
                 },
             },
             {
-                headerName: 'F',
-                width: 60,
-                cellStyle: { align: 'center' },
+                headerName: 'E',
                 cellRenderer: (params: any) => {
                     if (!params.data.goalData || !params.data.goalData[0]) return '';
                     const goalData = params.data.goalData[0];
-                    const goal: ICharacterAscendGoal = getAscensionGoal(goalData.goalId)!;
-                    return <StarsIcon stars={goal.starsEnd} />;
+                    return (
+                        <span>
+                            costs {goalData.goalCost} <MiscIcon icon={'energy'} height={15} width={15} />
+                        </span>
+                    );
                 },
             },
         ];
     }
-
-    const [goalDefs, setGoalDefs] = useState(getColumnDefs());
 
     return (
         <AgGridReact
