+++ src/fsd/1-pages/plan-campaign-progression/campaign-progression.route.tsx
