+++ src/fsd/1-pages/plan-campaign-progression/index.ts
@@ -1,0 +1,2 @@
+export { campaignProgressionLazyRoute } from './campaign-progression.route';
+export { campaignProgressionMenuItem } from './campaign-progression.menu-item';
