+++ src/fsd/4-entities/character/@x/goal.ts
@@ -1,0 +1,1 @@
+export type { IRankLookup } from '../model';
