+++ src/fsd/4-entities/goal/enums.ts
@@ -1,0 +1,14 @@
+export enum PersonalGoalType {
+    None = 0,
+    UpgradeRank = 1,
+    Ascend = 2,
+    Unlock = 3,
+    MowAbilities = 4,
+    CharacterAbilities = 5,
+}
+
+export enum CampaignsLocationsUsage {
+    None = 0,
+    BestTime = 1,
+    LeastEnergy = 2,
+}
