+++ src/fsd/4-entities/goal/index.ts
@@ -1,0 +1,8 @@
+export { PersonalGoalType, CampaignsLocationsUsage } from './enums';
+export type {
+    ICharacterUpgradeRankGoal,
+    ICharacterUpgradeMow,
+    ICharacterUnlockGoal,
+    ICharacterAscendGoal,
+    ICharacterRaidGoalSelectBase,
+} from './model';
