+++ src/fsd/4-entities/upgrade/index.ts
@@ -8,6 +8,8 @@ export type {
     IMaterial,
     IMaterialRecipeIngredient,
     IMaterialEstimated2,
+    IRecipeExpandedUpgrade,
+    IRecipeExpandedUpgradeData,
 } from './model';
 export { UpgradeImage } from './upgrade-image';
 export { UpgradeControl } from './upgrade-control';
