+++ src/models/enums.ts
@@ -6,21 +6,6 @@
     VeryHard,
 }
 
-export enum PersonalGoalType {
-    None = 0,
-    UpgradeRank = 1,
-    Ascend = 2,
-    Unlock = 3,
-    MowAbilities = 4,
-    CharacterAbilities = 5,
-}
-
-export enum CampaignsLocationsUsage {
-    None = 0,
-    BestTime = 1,
-    LeastEnergy = 2,
-}
-
 export enum DailyRaidsStrategy {
     leastEnergy,
     leastTime,
@@ -30,3 +15,4 @@ export enum DailyRaidsStrategy {
 
 // Re-export from FSD entities
 export { Campaign, CampaignType } from '@/fsd/4-entities/campaign';
+export { PersonalGoalType, CampaignsLocationsUsage } from '@/fsd/4-entities/goal';
