+++ src/v2/features/goals/goals.models.ts
@@ -6,9 +6,14 @@ import {
     IDailyRaidsPreferences,
 } from 'src/models/interfaces';
 
-import { Alliance, Faction, Rank, Rarity, RarityStars } from '@/fsd/5-shared/model';
-
-import { IRankLookup, IUnitUpgradeRank } from '@/fsd/4-entities/character';
+import { IUnitUpgradeRank } from '@/fsd/4-entities/character';
+import {
+    ICharacterAscendGoal,
+    ICharacterRaidGoalSelectBase,
+    ICharacterUnlockGoal,
+    ICharacterUpgradeMow,
+    ICharacterUpgradeRankGoal,
+} from '@/fsd/4-entities/goal';
 import { IBaseUpgrade } from '@/fsd/4-entities/upgrade';
 
 import { ICharacterAbilitiesMaterialsTotal, IXpEstimate } from 'src/v2/features/characters/characters.models';
@@ -20,40 +25,6 @@ export type CharacterRaidGoalSelect =
     | ICharacterUnlockGoal
     | ICharacterUpgradeMow
     | ICharacterUpgradeAbilities;
-
-export interface ICharacterRaidGoalSelectBase {
-    priority: number;
-    include: boolean;
-    goalId: string;
-    unitId: string;
-    unitName: string;
-    unitIcon: string;
-    unitAlliance: Alliance;
-    notes: string;
-}
-
-export interface ICharacterUpgradeRankGoal extends ICharacterRaidGoalSelectBase, IRankLookup {
-    type: PersonalGoalType.UpgradeRank;
-
-    rarity: Rarity;
-    level: number;
-    xp: number;
-}
-
-export interface ICharacterUpgradeMow extends ICharacterRaidGoalSelectBase {
-    type: PersonalGoalType.MowAbilities;
-
-    primaryStart: number;
-    primaryEnd: number;
-
-    secondaryStart: number;
-    secondaryEnd: number;
-    upgradesRarity: Rarity[];
-
-    shards: number;
-    stars: RarityStars;
-    rarity: Rarity;
-}
 
 export interface ICharacterUpgradeAbilities extends ICharacterRaidGoalSelectBase {
     type: PersonalGoalType.CharacterAbilities;
@@ -79,28 +50,6 @@ export interface IGoalEstimate {
     xpEstimateAbilities?: IXpEstimate;
     mowEstimate?: IMowMaterialsTotal;
     abilitiesEstimate?: ICharacterAbilitiesMaterialsTotal;
-}
-
-export interface ICharacterUnlockGoal extends ICharacterRaidGoalSelectBase {
-    type: PersonalGoalType.Unlock;
-
-    shards: number;
-    rank: Rank;
-    rarity: Rarity;
-    faction: Faction;
-    campaignsUsage: CampaignsLocationsUsage;
-}
-
-export interface ICharacterAscendGoal extends ICharacterRaidGoalSelectBase {
-    type: PersonalGoalType.Ascend;
-
-    rarityStart: Rarity;
-    starsStart: RarityStars;
-    starsEnd: RarityStars;
-    rarityEnd: Rarity;
-    shards: number;
-    onslaughtShards: number;
-    campaignsUsage: CampaignsLocationsUsage;
 }
 
 export interface IEstimatedAscensionSettings {
@@ -216,20 +165,10 @@ export interface ICharacterUpgradeEstimate extends IBaseUpgrade {
     isFinished: boolean;
 }
 
-/**
- * Holds the fully-expanded recipe for an upgrade material. One can then
- * reference the necessary IBaseUpgrade objects to get the full details.
- * That is, to say, expandedRecipe contains only uncraftable materials. If
- * this material is already uncraftable, expandedRecipe is empty.
- */
-export interface IRecipeExpandedUpgrade {
-    id: string;
-    label: string;
-    rarity: Rarity;
-    iconPath: string;
-    expandedRecipe: Record<string, number>;
-    crafted: boolean;
-    stat: string | 'Health' | 'Damage' | 'Armour' | 'Shard';
-}
-
-export type IRecipeExpandedUpgradeData = Record<string, IRecipeExpandedUpgrade>;
+export type {
+    ICharacterAscendGoal,
+    ICharacterRaidGoalSelectBase,
+    ICharacterUnlockGoal,
+    ICharacterUpgradeMow,
+    ICharacterUpgradeRankGoal,
+};
